Add files via upload
As a person who likes to learn and does not mind reading through the big programming books that seem to appear with every revision and release of the latest version of some technology.  I present my Microsoft Certified Professional Transcript, where you will see a history of the Microsoft exams I have taken as my career demanded.  I believe technology exams are worth doing.  Yes, experience is important, but exams coupled with developers who are willing to read up on and learn new technologies, I think makes for the best professionalism in the industry.